requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# 前端界面
gradio==4.8.0

# 深度学习框架（最新版本）
# 注意：RTX 4070显卡需要CUDA 12.1版本的PyTorch！
# 手动安装命令：pip install torch==2.1.2+cu121 torchvision==0.16.2+cu121 torchaudio==2.1.2+cu121 --index-url https://download.pytorch.org/whl/cu121
# torch==2.1.2+cu121
# torchvision==0.16.2+cu121  
# torchaudio==2.1.2+cu121
pytorch-lightning==2.1.3

# 音频处理
librosa==0.10.1
soundfile==0.12.1
sounddevice==0.4.6
pyaudio==0.2.13
webrtcvad==2.0.10

# 声纹识别训练专用
speechbrain==0.5.16
resemblyzer==0.1.4
nemo-toolkit[asr]==1.22.0
asteroid==0.7.0

# 自学习和在线学习
avalanche-lib==0.5.0
continual-learning==0.1.0

# AI/ML库（训练优先，兼容性版本）
transformers==4.36.0  # 保持稳定版本，与所有训练库兼容
accelerate==1.9.0     # 最新版本，更好的训练加速性能
datasets==2.16.1      # 兼容transformers 4.36.0
numpy>=1.22,<1.24     # 兼容nemo-toolkit等训练库的关键约束
scipy==1.11.4        # 兼容numpy<1.24的最高版本
scikit-learn==1.3.2   # 兼容numpy<1.24的最高版本

# 模型优化和加速（训练兼容版本）
onnx==1.16.2          # 较新版本，兼容numpy<1.24
# onnxruntime-gpu==1.16.3  # 暂时注释，需要numpy>=1.24.2，与训练库冲突
# tensorrt==9.2.0.post12.dev5  # 需要特殊安装：pip install --extra-index-url https://pypi.nvidia.com tensorrt
# triton==2.1.0  # 已通过torch内置

# 数据处理和增强
albumentations==1.3.1
audiomentations==0.34.1
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4

# 实时音频处理
pyaudio==0.2.13
sounddevice==0.4.6
# rtaudio==0.2.0  # Windows不支持
resampy==0.4.2

# 模型训练工具
wandb==0.16.1
tensorboard==2.15.1
hydra-core==1.3.2
omegaconf==2.3.0

# 语音相关（扩展）
faster-whisper==0.10.0
coqui-TTS==0.24.2
# espnet==202310  # 版本可能过旧，暂时注释
# fairseq==0.12.2  # 版本可能过旧，暂时注释

# 工具库
pyyaml==6.0.1
requests==2.31.0
tqdm==4.66.1
click==8.1.7
rich==13.7.0

# 开发工具
pytest==7.4.3
black==23.11.0
isort==5.12.0
pre-commit==3.6.0
